Refactor user creation
Moving user creation to the hab install step.  This will allow it to be
skipped if the hab install is skipped.

Also moved from assuming useradd is present, to installing busybox and
using the included adduser.  We're already pulling packages from the
depot, so this seems like a sane thing to do.

Moved user install to its own function.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>