Refactoring of the user TOML inner workings
Moved from using the communicator's file upload method.  The upload
method required giving extra access to the `/hab/svc/<name>` dir to
allow uploading.  We're now just echo'ing out to tee so that we can
drop the file off in the right place whether we're root or not by using
sudo.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>