Adding Bind struct and bind blocks
Added a struct to represent binds.  This allows binds to be specified
by either the `binds = ["alias:service.group"]` argument, or bind
blocks:

```
bind {
  alias = "alias"
  service = "service"
  group = "group"
}
```

Either format may be used, and used together in the same service
because we just take the strings from `binds` and turn them into Bind
structs when decoding the config.

I think this block format could be useful when setting part of the bind
based on a variable or resource reference, rather than doing string
interpolation.

```
bind {
  alias = "database"
  service = "mysql"
  group = "${var.environment}"
}
```

Signed-off-by: Nolan Davidson <ndavidson@chef.io>